internal/genv: fix release notes link for minor versions
Also corrected the code in case of unrecognized version.

According to the docs, nil is returned when the regex does not match.
And in case of a partial or full match, the length of the slice
is always equal to the no. of subexpressions.
For the subexpressions that did not match, the string will be empty.

So the length check was logically wrong. It's just that since the length
of a nil slice is 0, it accidentally turned out to be correct.

Fixes golang/go#26976

Change-Id: Ic9e43d13877d747df20c15949436e355d2fd448d
Reviewed-on: https://go-review.googlesource.com/c/160920
Reviewed-by: Filippo Valsorda <filippo@golang.org>