gotip: add $GOROOT/bin to $PATH before starting go binary
This change applies the fix from CL 143545 to the gotip command,
which does not use the golang.org/dl/internal/version package at
this time.

Copy the necessary dependencies for now. In the future, the API
of ./internal/version can be reworked to make it possible for
gotip and all other commands to share more code.

Also remove a blank line in ./internal/version code, because it was
breaking up a single logical block of code and reducing readability.

Fixes golang/go#35507

Change-Id: Ic10e561cccee5259673c76ad679029c68641615f
Reviewed-on: https://go-review.googlesource.com/c/dl/+/206537
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>