Do not rebuild gotip if the HEAD does not change
On successful build a sentinel zero-byte file is written,
similar to .unpacked-success in internal/version. If such
file exists on next gotip download, compare HEADs before
and after FETCH_HEAD checkout, and trigger the build iff
they differ. In that case, sentinel file is removed to
make git clean happy and indicate that rebuild is needed.