gotip: don't assume origin as the git remote name
When running "gotip download", we try to update the gotip working tree
by running several git commands with the hard-coded name "origin". The
name "origin" is the default in git, but this is configurable, so it's
not necessarily safe to assume that a remote named "origin" will always
exist.

We can, however, force the name "origin" when doing the initial clone,
so that later git operations that interact with the remote can safely
assume that the name "origin" exists. This isn't totally reliable,
because if a user manually renames the remote, the commands that assume
"origin" will still fail. This change, though, ensures that the default
configuration will work even if a user has set clone.defaultRemoteName
in their local git configuration.